Add SASS mixin: force-webkit-scrollbar
This can be handy when you *know* you want a scrollbar. On devices with
a touch screen (or a trackpad, such as a MacBook) scrollbars won't show up.
This is typically not an issue, unless it's important to indicate to the user that
there is more content, and it's not particularly evident.