Use custom post type content templates in single.php where appropriate.
Rather than just grabbing post formats, we are checking first if the post type = "post".
If it does, then go ahead and get_post_format(). But if not, then we'd like to use
the appropriate post type content template. I primarily am doing this to allow previewing
of custom post types without creating template files for something that likely won't be
shown to public view anyway.

Perhaps we should create a mechanism that will only show these pages to logged-in
viewers. Probably not right now.