Add toggle for featured images to indicate they have a light background
... and styling for the header overlaying it to accommodate.

By adding a checkbox custom field directly into the featured image metabox,
we can indicate whether the image we've just added is too light for our white
heading to look good on top of it. Then we add a filter to the post_class hook
and we've got something to style. Voilà!