SCSS: Move .toggles menu styling into a module; hide toggle if not in use.
Goal here, though not utilized yet, is to facilitate the use of the toggles concept for certain media queries only, so we could have .m-toggle, .t-toggle, and d-toggle to use them only when appropriate.

I am not yet convinced this approach will bear fruit. And right now the toggles are not being used. But this may come in handy down the road.